feat: full commonJS compatibility and webpack build
+ to handle algoliaBundle
+ fixed some angular bugs that where not visible since we always had
jquery in the page. Angular alone build was not working

Angular build is still not working when used without jquery in the
page. Something is still wrong, there are differences between jqLite
and jQuery.

Two tests are broken, cannot find why. Some tests were broken and
relied on other data.