fix(postMessage): avoid using postMessage when feasible
Before this commit, we were using setImmediate which was then
polyfilled by webpack in a way that is not efficient
and led to issues like #142 where postMessage was used
no matter what.

Now we use the `immediate` dependency which will first try
to use nextTick, then mutation observers (IE11) then fallback
to postMessage but in a way that should not bug any user
registering for global messages.

fixes #142